InfoTrigger 상속 구조 변경
