--- Assets/02.Scripts/Core/Trigger/GetInformationTrigger.cs
@@ -1,34 +1,0 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.EventSystems;
-using UnityEngine.UI;
-
-public class GetInformationTrigger : InformationTrigger
-{
-    protected override void OnEnable()
-    {
-        base.OnEnable();
-    }
-
-    protected override void Start()
-    {
-        base.Start();
-    }
-
-    public override void OnPointerClick(PointerEventData eventData)
-    {
-        base.OnPointerClick(eventData);
-    }
-
-    public override void OnPointerEnter(PointerEventData eventData)
-    {
-        base.OnPointerEnter(eventData);
-    }
-
-    public override void OnPointerExit(PointerEventData eventData)
-    {
-        base.OnPointerExit(eventData);
-    }
-}
