+++ Assets/03.Prefabs/ImageViewers/ImageViewer/Body/InvestigationBody.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 4709044759897787283}
   - component: {fileID: 1257813706623605249}
   - component: {fileID: 7360983704865336105}
-  - component: {fileID: 2423775368447094597}
   m_Layer: 5
   m_Name: Image
   m_TagString: Untagged
@@ -77,22 +76,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &2423775368447094597
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1210932557179886032}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 5
-  information: MurderFirstPlan
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &7418697032215131764
 GameObject:
   m_ObjectHideFlags: 0
