+++ Assets/03.Prefabs/ImageViewers/ImageViewer/Body/Pet/FirstMetPetsBody.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 5466862493306471427}
   - component: {fileID: 5087769353586198298}
   - component: {fileID: 7018996266200689560}
-  - component: {fileID: 3928711643622987199}
   m_Layer: 5
   m_Name: PetBreedTrigger
   m_TagString: Untagged
@@ -77,21 +76,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &3928711643622987199
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7969038130583203241}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 4
-  information: PetBreed
-  needStringList: []
 --- !u!1 &8454321266748870906
 GameObject:
   m_ObjectHideFlags: 0
@@ -186,6 +170,7 @@ MonoBehaviour:
   defaultScale: 0.6
   maxImageScale: 3
   zoomSpeed: 0.1
+  isDiscord: 0
 --- !u!114 &8051523755304602131
 MonoBehaviour:
   m_ObjectHideFlags: 0
