+++ Assets/03.Prefabs/ImageViewers/ImageViewer/Body/Pet/PetsReceiptBody.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 8113674711046166770}
   - component: {fileID: 3918721558804777465}
   - component: {fileID: 7682459060459678694}
-  - component: {fileID: 4767481272861950408}
   m_Layer: 5
   m_Name: SuspectNameTrigger (1)
   m_TagString: Untagged
@@ -77,22 +76,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &4767481272861950408
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2425534920355660855}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 1
-  information: SuspectName
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &6354995909807982097
 GameObject:
   m_ObjectHideFlags: 0
@@ -104,7 +87,6 @@ GameObject:
   - component: {fileID: 2347512145256222959}
   - component: {fileID: 1212587694329310416}
   - component: {fileID: 6582059967626977734}
-  - component: {fileID: 6572416611722701068}
   m_Layer: 5
   m_Name: PetAgeTrigger (1)
   m_TagString: Untagged
@@ -170,22 +152,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &6572416611722701068
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6354995909807982097}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 4
-  information: PetAge
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &6423995700564304842
 GameObject:
   m_ObjectHideFlags: 0
@@ -197,7 +163,6 @@ GameObject:
   - component: {fileID: 9213069681214945222}
   - component: {fileID: 758799134667563683}
   - component: {fileID: 5328895276918163002}
-  - component: {fileID: 9014954476658218671}
   m_Layer: 5
   m_Name: SuspectNameTrigger
   m_TagString: Untagged
@@ -263,22 +228,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &9014954476658218671
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6423995700564304842}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 1
-  information: SuspectName
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &7969038130583203241
 GameObject:
   m_ObjectHideFlags: 0
@@ -290,7 +239,6 @@ GameObject:
   - component: {fileID: 5466862493306471427}
   - component: {fileID: 5087769353586198298}
   - component: {fileID: 7018996266200689560}
-  - component: {fileID: 3928711643622987199}
   m_Layer: 5
   m_Name: PetAgeTrigger
   m_TagString: Untagged
@@ -356,22 +304,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &3928711643622987199
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7969038130583203241}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 4
-  information: PetAge
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &8454321266748870906
 GameObject:
   m_ObjectHideFlags: 0
