+++ Assets/03.Prefabs/ImageViewers/ImageViewer/Body/Suspect/SuspectBody.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 7652366894372134542}
   - component: {fileID: 2718375608617401958}
   - component: {fileID: 6497852548306768210}
-  - component: {fileID: 6413235127758945494}
   m_Layer: 5
   m_Name: SuspectRelationWithVictimTrigger
   m_TagString: Untagged
@@ -77,22 +76,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &6413235127758945494
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4164780261796539905}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 1
-  information: SuspectRelationWithVictim
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &5557635916138299879
 GameObject:
   m_ObjectHideFlags: 0
@@ -104,7 +87,6 @@ GameObject:
   - component: {fileID: 4802156787450827666}
   - component: {fileID: 8635529032692569986}
   - component: {fileID: 2230448445830546692}
-  - component: {fileID: 931781870933853748}
   m_Layer: 5
   m_Name: SuspectIsLivingWithVictimTrigger
   m_TagString: Untagged
@@ -170,22 +152,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &931781870933853748
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5557635916138299879}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 1
-  information: SuspectIsLivingWithVictim
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &6302246490556891876
 GameObject:
   m_ObjectHideFlags: 0
@@ -197,7 +163,6 @@ GameObject:
   - component: {fileID: 2963885122708605576}
   - component: {fileID: 3572264636680241334}
   - component: {fileID: 1057447696204717188}
-  - component: {fileID: 1491586440960380519}
   m_Layer: 5
   m_Name: VictimNameTrigger_2
   m_TagString: Untagged
@@ -263,22 +228,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &1491586440960380519
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6302246490556891876}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 3
-  information: VictimName
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &7969038130583203241
 GameObject:
   m_ObjectHideFlags: 0
@@ -290,7 +239,6 @@ GameObject:
   - component: {fileID: 5466862493306471427}
   - component: {fileID: 5087769353586198298}
   - component: {fileID: 7018996266200689560}
-  - component: {fileID: 3928711643622987199}
   m_Layer: 5
   m_Name: SuspectNameTrigger
   m_TagString: Untagged
@@ -356,22 +304,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &3928711643622987199
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7969038130583203241}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 1
-  information: SuspectName
-  needInformaitonList: []
-  backgroundImage: {fileID: 7018996266200689560}
 --- !u!1 &8004805431077693257
 GameObject:
   m_ObjectHideFlags: 0
@@ -383,7 +315,6 @@ GameObject:
   - component: {fileID: 4327594759877386554}
   - component: {fileID: 3700819356787269653}
   - component: {fileID: 3085786317230903967}
-  - component: {fileID: 532809444547461868}
   m_Layer: 5
   m_Name: VictimNameTrigger_1
   m_TagString: Untagged
@@ -449,22 +380,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &532809444547461868
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8004805431077693257}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 3
-  information: VictimName
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &8096703127126436895
 GameObject:
   m_ObjectHideFlags: 0
@@ -476,7 +391,6 @@ GameObject:
   - component: {fileID: 237632974583518270}
   - component: {fileID: 8546605852559308005}
   - component: {fileID: 6328556814298540932}
-  - component: {fileID: 5922680263709713838}
   m_Layer: 5
   m_Name: SuspectResidenceTrigger
   m_TagString: Untagged
@@ -542,23 +456,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &5922680263709713838
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8096703127126436895}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 1
-  information: SuspectResidence
-  needInformaitonList:
-  - {fileID: 11400000, guid: 6811cb31453366f488ec25b3104c3792, type: 2}
-  backgroundImage: {fileID: 0}
 --- !u!1 &8454321266748870906
 GameObject:
   m_ObjectHideFlags: 0
@@ -658,6 +555,7 @@ MonoBehaviour:
   defaultScale: 2
   maxImageScale: 4
   zoomSpeed: 0.1
+  isDiscord: 0
 --- !u!114 &8051523755304602131
 MonoBehaviour:
   m_ObjectHideFlags: 0
