+++ Assets/03.Prefabs/ImageViewers/ImageViewer/Body/University/UniversityEntrance.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 6851687764406688465}
   - component: {fileID: 8192837358333839770}
   - component: {fileID: 1251341221667891102}
-  - component: {fileID: 8499753100427420244}
   m_Layer: 5
   m_Name: Image (2)
   m_TagString: Untagged
@@ -77,25 +76,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &8499753100427420244
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 56174166182297002}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 3
-  information: VictimJob
-  needInformaitonList:
-  - {fileID: 11400000, guid: 523a26259b6e5c34db4ea04f5c6e1edb, type: 2}
-  backgroundImage: {fileID: 0}
-  monologType: 0
-  delay: 0
 --- !u!1 &510609581959584136
 GameObject:
   m_ObjectHideFlags: 0
@@ -107,7 +87,6 @@ GameObject:
   - component: {fileID: 3319949466986936470}
   - component: {fileID: 8882266176595402814}
   - component: {fileID: 4100066818824722126}
-  - component: {fileID: 6946313584576756745}
   m_Layer: 5
   m_Name: fakeImage
   m_TagString: Untagged
@@ -173,24 +152,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &6946313584576756745
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 510609581959584136}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 3
-  information: VictimName
-  needInformaitonList: []
-  backgroundImage: {fileID: 4100066818824722126}
-  monologType: 3
-  delay: 0.5
 --- !u!1 &8051555751420258238
 GameObject:
   m_ObjectHideFlags: 0
@@ -202,7 +163,6 @@ GameObject:
   - component: {fileID: 5510239587169855654}
   - component: {fileID: 5567022826978804293}
   - component: {fileID: 327622887133995944}
-  - component: {fileID: 8266871380267591040}
   m_Layer: 5
   m_Name: Image (1)
   m_TagString: Untagged
@@ -268,25 +228,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &8266871380267591040
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8051555751420258238}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 3
-  information: VictimUniversity
-  needInformaitonList:
-  - {fileID: 11400000, guid: 523a26259b6e5c34db4ea04f5c6e1edb, type: 2}
-  backgroundImage: {fileID: 0}
-  monologType: 0
-  delay: 0
 --- !u!1 &8454321266748870906
 GameObject:
   m_ObjectHideFlags: 0
@@ -408,7 +349,6 @@ GameObject:
   - component: {fileID: 4972097902862993075}
   - component: {fileID: 8792300955058653062}
   - component: {fileID: 6258377092662166584}
-  - component: {fileID: 3004462910088575013}
   m_Layer: 5
   m_Name: Image
   m_TagString: Untagged
@@ -474,22 +414,3 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &3004462910088575013
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8606558563041626316}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 3
-  information: VictimUniversity
-  needInformaitonList:
-  - {fileID: 11400000, guid: 523a26259b6e5c34db4ea04f5c6e1edb, type: 2}
-  backgroundImage: {fileID: 0}
-  monologType: 0
-  delay: 0
