+++ Assets/03.Prefabs/Mails/OrderListMail.prefab
@@ -1015,7 +1015,6 @@ GameObject:
   - component: {fileID: 6379765225556304249}
   - component: {fileID: 3412548211133424875}
   - component: {fileID: 7804691503847622347}
-  - component: {fileID: 3244293957569172125}
   m_Layer: 5
   m_Name: SuspectNameTrigger (3)
   m_TagString: Untagged
@@ -1081,22 +1080,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &3244293957569172125
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 648112311804687600}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 1
-  information: SuspectName
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &1031586411134806365
 GameObject:
   m_ObjectHideFlags: 0
@@ -1108,7 +1091,6 @@ GameObject:
   - component: {fileID: 3950760783471291352}
   - component: {fileID: 3058896587248591791}
   - component: {fileID: 7962703546693469651}
-  - component: {fileID: 3583648595752068842}
   m_Layer: 5
   m_Name: SuspectSpotTrigger (3)
   m_TagString: Untagged
@@ -1174,22 +1156,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &3583648595752068842
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1031586411134806365}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 1
-  information: SuspectResidence
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &1238219664381400003
 GameObject:
   m_ObjectHideFlags: 0
@@ -2225,7 +2191,6 @@ GameObject:
   - component: {fileID: 8244999888411720362}
   - component: {fileID: 1178053758703348242}
   - component: {fileID: 8597521983880397787}
-  - component: {fileID: 1081959840059247733}
   m_Layer: 5
   m_Name: SuspectNameTrigger (2)
   m_TagString: Untagged
@@ -2291,22 +2256,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &1081959840059247733
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2389176432179713569}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 1
-  information: SuspectName
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &2502312714092071607
 GameObject:
   m_ObjectHideFlags: 0
@@ -2669,7 +2618,6 @@ GameObject:
   - component: {fileID: 5839161523323836132}
   - component: {fileID: 6177919390402176787}
   - component: {fileID: 902964576716534205}
-  - component: {fileID: 2413463722231947444}
   m_Layer: 5
   m_Name: SuspectNameTrigger (1)
   m_TagString: Untagged
@@ -2735,22 +2683,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &2413463722231947444
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2627220810019841236}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 1
-  information: SuspectName
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &2650708875915959304
 GameObject:
   m_ObjectHideFlags: 0
@@ -3344,7 +3276,6 @@ GameObject:
   - component: {fileID: 5593269609840700039}
   - component: {fileID: 8635282741072864943}
   - component: {fileID: 816761199319296708}
-  - component: {fileID: 9030996306700913297}
   m_Layer: 5
   m_Name: SuspectPhonNumberTrigger
   m_TagString: Untagged
@@ -3410,22 +3341,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &9030996306700913297
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3784160109195976588}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 7
-  information: SuspectPhoneNum
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &3880468333578578800
 GameObject:
   m_ObjectHideFlags: 0
@@ -4377,7 +4292,6 @@ GameObject:
   - component: {fileID: 5276393747079847019}
   - component: {fileID: 1240039159524816572}
   - component: {fileID: 3319503882873996444}
-  - component: {fileID: 1458310212162766916}
   m_Layer: 5
   m_Name: KnifeInfoTrigger
   m_TagString: Untagged
@@ -4443,22 +4357,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &1458310212162766916
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5819103913260146423}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 5
-  information: MurderStuff
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &5870789226403920731
 GameObject:
   m_ObjectHideFlags: 0
@@ -7233,7 +7131,6 @@ GameObject:
   - component: {fileID: 5662584846336484452}
   - component: {fileID: 2568401012481188351}
   - component: {fileID: 545469340568840107}
-  - component: {fileID: 765754342810279938}
   m_Layer: 5
   m_Name: SuspectNameTrigger
   m_TagString: Untagged
@@ -7299,22 +7196,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &765754342810279938
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 9109429293133118970}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 1
-  information: SuspectName
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1001 &681491292448221469
 PrefabInstance:
   m_ObjectHideFlags: 0
