+++ Assets/03.Prefabs/Mails/ReplyCCTVMail.prefab
@@ -275,7 +275,7 @@ MonoBehaviour:
   m_HandleRect: {fileID: 8423899337503559244}
   m_Direction: 2
   m_Value: 0
-  m_Size: 1
+  m_Size: 0.33475
   m_NumberOfSteps: 0
   m_OnValueChanged:
     m_PersistentCalls:
@@ -1603,7 +1603,6 @@ GameObject:
   - component: {fileID: 3608363669756509058}
   - component: {fileID: 6816480145207052596}
   - component: {fileID: 33339143865701537}
-  - component: {fileID: 4291606091569885918}
   - component: {fileID: 8091426527289616764}
   m_Layer: 5
   m_Name: CCTVImage
@@ -1671,22 +1670,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &4291606091569885918
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4401418421146838489}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 5
-  information: MurderrFirstConditions
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!114 &8091426527289616764
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -2133,7 +2116,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 1, y: 1}
-  m_AnchoredPosition: {x: 0, y: 0}
+  m_AnchoredPosition: {x: 0, y: 532.19995}
   m_SizeDelta: {x: 0, y: 800}
   m_Pivot: {x: 0.5, y: 1}
 --- !u!114 &7690956104554811465
