+++ Assets/03.Prefabs/Mails/RequestMail.prefab
@@ -342,8 +342,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 890, y: -80}
-  m_SizeDelta: {x: 1780, y: 60}
+  m_AnchoredPosition: {x: 208, y: -80}
+  m_SizeDelta: {x: 416, y: 60}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &6067498730496659133
 CanvasRenderer:
@@ -556,8 +556,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 890, y: -111}
-  m_SizeDelta: {x: 1780, y: 2}
+  m_AnchoredPosition: {x: 208, y: -111}
+  m_SizeDelta: {x: 416, y: 2}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &8392622637645722308
 CanvasRenderer:
@@ -766,7 +766,6 @@ GameObject:
   - component: {fileID: 1195841668969878344}
   - component: {fileID: 4925964721947272424}
   - component: {fileID: 23511999669505627}
-  - component: {fileID: 5276137182067570045}
   m_Layer: 5
   m_Name: MurderSecondPlanTrigger
   m_TagString: Untagged
@@ -832,22 +831,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &5276137182067570045
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3355902412845527259}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 5
-  information: MurderSecondPlan
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &3486348405228900804
 GameObject:
   m_ObjectHideFlags: 0
@@ -883,8 +866,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 890, y: -25}
-  m_SizeDelta: {x: 1780, y: 50}
+  m_AnchoredPosition: {x: 208, y: -25}
+  m_SizeDelta: {x: 416, y: 50}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &301622890752720043
 CanvasRenderer:
@@ -1456,8 +1439,8 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 7159872411351427328}
   m_HandleRect: {fileID: 6548333098288591774}
   m_Direction: 2
-  m_Value: 0
-  m_Size: 1
+  m_Value: 0.9999972
+  m_Size: 0.8381472
   m_NumberOfSteps: 0
   m_OnValueChanged:
     m_PersistentCalls:
@@ -1587,8 +1570,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 890, y: -112}
-  m_SizeDelta: {x: 1780, y: 159.79}
+  m_AnchoredPosition: {x: 208, y: -112}
+  m_SizeDelta: {x: 416, y: 159.79}
   m_Pivot: {x: 0.5, y: 1}
 --- !u!222 &6521736088717377727
 CanvasRenderer:
