+++ Assets/03.Prefabs/Notepad/NotepadBody/ClientInformation.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 8523147019716311816}
   - component: {fileID: 8713234557605341815}
   - component: {fileID: 7975166936551050293}
-  - component: {fileID: 5735439494749884656}
   m_Layer: 5
   m_Name: OwnerName
   m_TagString: Untagged
@@ -77,21 +76,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &5735439494749884656
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1498654109712719221}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 0
-  information: OwnerName
-  needInformaitonList: []
 --- !u!1 &2216837552259616109
 GameObject:
   m_ObjectHideFlags: 0
@@ -103,7 +87,6 @@ GameObject:
   - component: {fileID: 8304652845122886565}
   - component: {fileID: 4679198721457762855}
   - component: {fileID: 7687234239098230156}
-  - component: {fileID: 6344248457757130810}
   m_Layer: 5
   m_Name: OwnerBirth
   m_TagString: Untagged
@@ -169,21 +152,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &6344248457757130810
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2216837552259616109}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 0
-  information: OwnerBirth
-  needInformaitonList: []
 --- !u!1 &3863809119840419671
 GameObject:
   m_ObjectHideFlags: 0
@@ -195,7 +163,6 @@ GameObject:
   - component: {fileID: 8388628148494137275}
   - component: {fileID: 7063900191105217245}
   - component: {fileID: 1497452161395605955}
-  - component: {fileID: 7589202617849403349}
   m_Layer: 5
   m_Name: OwnerGender
   m_TagString: Untagged
@@ -261,21 +228,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &7589202617849403349
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3863809119840419671}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 0
-  information: OwnerGender
-  needInformaitonList: []
 --- !u!1 &6760515456448002866
 GameObject:
   m_ObjectHideFlags: 0
