+++ Assets/03.Prefabs/Notepad/NotepadBody/ZooglePasswordBody.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 2292637119681074177}
   - component: {fileID: 4630136446240326244}
   - component: {fileID: 7805428471055824193}
-  - component: {fileID: 2158136916599922196}
   m_Layer: 5
   m_Name: PasswordTrigger
   m_TagString: Untagged
@@ -77,21 +76,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &2158136916599922196
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1813979691788382910}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 7
-  information: ZooglePassword
-  needInformaitonList: []
 --- !u!1 &6760515456448002866
 GameObject:
   m_ObjectHideFlags: 0
