+++ Assets/03.Prefabs/UI/WindowIcon/IconRightClick/PropertyBody/FirstMetpetPropertyBody.prefab
@@ -1778,7 +1778,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 7529898708202948301}
-  - component: {fileID: 7041848844038327625}
   - component: {fileID: 3411450884478910236}
   - component: {fileID: 3346758250788373384}
   m_Layer: 0
@@ -1808,22 +1807,6 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0.5}
   m_SizeDelta: {x: 10, y: 5}
   m_Pivot: {x: 0.5, y: 0.5}
---- !u!114 &7041848844038327625
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8820502937905930641}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 4
-  information: PetAdoptionDate
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!222 &3411450884478910236
 CanvasRenderer:
   m_ObjectHideFlags: 0
