+++ Assets/03.Prefabs/UI/Windows/Browsers/Sites/Home/HomeSite.prefab
@@ -2332,7 +2332,6 @@ GameObject:
   - component: {fileID: 4508488926433881160}
   - component: {fileID: 4670703011094837828}
   - component: {fileID: 6301908020071863486}
-  - component: {fileID: 7744325495514566334}
   m_Layer: 5
   m_Name: MurderRecord
   m_TagString: Untagged
@@ -2400,22 +2399,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &7744325495514566334
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3848889033483938854}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 821eea7ff19ed014084f76f8ce9d623d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  category: 5
-  information: MurderSecondConditions
-  needInformaitonList: []
-  backgroundImage: {fileID: 0}
 --- !u!1 &3902436176738446653
 GameObject:
   m_ObjectHideFlags: 0
